+++ Finder.go
@@ -36,7 +36,8 @@ func NewFinder() *Finder {
 	finder := Finder{}
 	finder.SelectTotalCount = true
 	finder.InjectionCheck = true
-	finder.values = make([]interface{}, 0)
+	//slice扩容会生成新的slice,最后要值复制接收.问:为什么cap是3?答:经验
+	finder.values = make([]interface{}, 0, 3)
 	return &finder
 }
 
@@ -88,9 +89,9 @@ func NewDeleteFinder(tableName string) *Finder {
 //Only splice SQL, E.g: finder.Append(" and name=123 ")
 func (finder *Finder) Append(s string, values ...interface{}) *Finder {
 
-	//不要自己构建finder,使用Newxxx方法
-	//Don't build finder by yourself, use Newxxx method
-	if finder.values == nil {
+	//不要自己构建finder,使用NewFinder()方法
+	//Don't build finder by yourself, use NewFinder() method
+	if finder == nil || finder.values == nil {
 		return nil
 	}
 
@@ -118,21 +119,24 @@ func (finder *Finder) Append(s string, values ...interface{}) *Finder {
 //AppendFinder 添加另一个Finder finder.AppendFinder(f)
 //AppendFinder Add another Finder . finder.AppendFinder(f)
 func (finder *Finder) AppendFinder(f *Finder) (*Finder, error) {
+	if finder == nil {
+		return finder, errors.New("finder-->AppendFinder()finder对象为nil")
+	}
 	if f == nil {
-		return nil, errors.New("finder-->AppendFinder参数是nil")
+		return finder, errors.New("finder-->AppendFinder()参数是nil")
 	}
 
-	//不要自己构建finder,使用Newxxx方法
-	//Don't build finder by yourself, use Newxxx method
+	//不要自己构建finder,使用NewFinder()方法
+	//Don't build finder by yourself, use NewFinder() method
 	if finder.values == nil {
-		return nil, errors.New("finder-->AppendFinder不要自己构建finder,使用Newxxx方法")
+		return finder, errors.New("finder-->AppendFinder()不要自己构建finder,使用NewFinder()方法")
 	}
 
 	//添加f的SQL
-	//SQL to add f。
+	//SQL to add f
 	sqlstr, err := f.GetSQL()
 	if err != nil {
-		return nil, err
+		return finder, err
 	}
 	finder.sqlstr = ""
 	finder.sqlBuilder.WriteString(sqlstr)
@@ -145,10 +149,10 @@ func (finder *Finder) AppendFinder(f *Finder) (*Finder, error) {
 //GetSQL 返回Finder封装的SQL语句
 //GetSQL Return the SQL statement encapsulated by the Finder
 func (finder *Finder) GetSQL() (string, error) {
-	//不要自己构建finder,使用Newxxx方法
-	//Don't build finder by yourself, use Newxxx method
-	if finder.values == nil {
-		return "", errors.New("finder-->GetSQL不要自己构建finder,使用Newxxx方法")
+	//不要自己构建finder,使用NewFinder方法
+	//Don't build finder by yourself, use NewFinder method
+	if finder == nil || finder.values == nil {
+		return "", errors.New("finder-->GetSQL()不要自己构建finder,使用NewFinder()方法")
 	}
 	if len(finder.sqlstr) > 0 {
 		return finder.sqlstr, nil
@@ -158,7 +162,7 @@ func (finder *Finder) GetSQL() (string, error) {
 	//包含单引号,属于非法字符串
 	//Contains single quotes, which are illegal strings
 	if finder.InjectionCheck && (strings.Contains(sqlstr, "'")) {
-		return "", errors.New("finder-->GetSQL SQL语句请不要直接拼接字符串参数!!!使用标准的占位符实现,例如  finder.Append(' and id=? and name=? ','123','abc')")
+		return "", errors.New("finder-->GetSQL()SQL语句请不要直接拼接字符串参数!!!使用标准的占位符实现,例如  finder.Append(' and id=? and name=? ','123','abc')")
 	}
 
 	//处理sql语句中的in,实际就是把数组变量展开,例如 id in(?) ["1","2","3"] 语句变更为 id in (?,?,?) 参数也展开到参数数组里
@@ -240,7 +244,7 @@ func (finder *Finder) GetSQL() (string, error) {
 		//数组类型的参数长度小于1,认为是有异常的参数
 		//The parameter length of the array type is less than 1, which is considered to be an abnormal parameter
 		if sliceLen < 1 {
-			return sqlstr, errors.New("finder-->GetSQL语句:" + sqlstr + ",第" + strconv.Itoa(i+1) + "个参数,类型是Array或者Slice,值的长度为0,请检查sql参数有效性")
+			return sqlstr, errors.New("finder-->GetSQL()语句:" + sqlstr + ",第" + strconv.Itoa(i+1) + "个参数,类型是Array或者Slice,值的长度为0,请检查sql参数有效性")
 		}
 
 		for j := 0; j < sliceLen; j++ {