+++ structFieldInfo.go
@@ -307,7 +307,7 @@ func getDBColumnFieldMap(typeOf *reflect.Type) (map[string]reflect.StructField, 
 func getDBColumnFieldNameSlice(typeOf *reflect.Type) ([]string, error) {
 	dbColumnFieldSlice, dbmapErr := getCacheStructFieldInfo(typeOf, dbColumnNameSlicePrefix)
 	if dbmapErr != nil {
-		return nil, errors.New("getDBColumnFieldNameSlice()-->getCacheStructFieldInfo()取值异常")
+		return nil, fmt.Errorf("->getDBColumnFieldNameSlice()-->getCacheStructFieldInfo()取值异常:%w", dbmapErr)
 	}
 	dbcfSlice, efOK := dbColumnFieldSlice.([]string)
 	if !efOK {
@@ -347,7 +347,7 @@ func getCacheStructFieldInfo(typeOf *reflect.Type, keyPrefix string) (interface{
 func getCacheStructFieldInfoMap(typeOf *reflect.Type, keyPrefix string) (map[string]reflect.StructField, error) {
 	dbColumnFieldMap, dbmapErr := getCacheStructFieldInfo(typeOf, keyPrefix)
 	if dbmapErr != nil {
-		return nil, errors.New("getCacheStructFieldInfoMap()-->getCacheStructFieldInfo()取值异常")
+		return nil, fmt.Errorf("->getCacheStructFieldInfoMap()-->getCacheStructFieldInfo()取值异常:%w", dbmapErr)
 	}
 	dbcfMap, efOK := dbColumnFieldMap.(map[string]reflect.StructField)
 	if !efOK {
@@ -540,7 +540,7 @@ func sqlRowsValues(ctx context.Context, rows *sql.Rows, driverValue *reflect.Val
 				typeOf := fieldValue.Type()
 				tempDriverValue, errGetDriverValue = converFunc.GetDriverValue(ctx, columnType, &typeOf, finder)
 				if errGetDriverValue != nil {
-					errGetDriverValue = fmt.Errorf("sqlRowsValues-->conver.GetDriverValue异常:%w", errGetDriverValue)
+					errGetDriverValue = fmt.Errorf("->sqlRowsValues-->conver.GetDriverValue异常:%w", errGetDriverValue)
 					FuncLogError(ctx, errGetDriverValue)
 					return errGetDriverValue
 				}
@@ -578,7 +578,7 @@ func sqlRowsValues(ctx context.Context, rows *sql.Rows, driverValue *reflect.Val
 		typeOf := fieldValue.Type()
 		rightValue, errConverDriverValue := driverValueInfo.converFunc.ConverDriverValue(ctx, driverValueInfo.columnType, &typeOf, driverValueInfo.tempDriverValue, finder)
 		if errConverDriverValue != nil {
-			errConverDriverValue = fmt.Errorf("sqlRowsValues-->conver.ConverDriverValue异常:%w", errConverDriverValue)
+			errConverDriverValue = fmt.Errorf("->sqlRowsValues-->conver.ConverDriverValue异常:%w", errConverDriverValue)
 			FuncLogError(ctx, errConverDriverValue)
 			return errConverDriverValue
 		}
@@ -614,7 +614,7 @@ func sqlRowsValuesFast(rows *sql.Rows, columns []string, dbColumnFieldMap map[st
 	//scan赋值.是一个指针数组,已经根据struct的属性类型初始化了,sql驱动能感知到参数类型,所以可以直接赋值给struct的指针.这样struct的属性就有值了
 	scanerr := rows.Scan(values...)
 	if scanerr != nil {
-		scanerr = fmt.Errorf("rows.Scan异常:%w", scanerr)
+		scanerr = fmt.Errorf("->rows.Scan异常:%w", scanerr)
 		FuncLogError(scanerr)
 		return scanerr
 	}